refactor(element3): move Avatar to src/components
