+++ packages/element3/packages/theme-chalk/src/index.scss
@@ -2,7 +2,6 @@
 @import "./alert.scss";
 @import "./aside.scss";
 @import "./autocomplete.scss";
-@import "./avatar.scss";
 @import "./backtop.scss";
 @import "./breadcrumb.scss";
 @import "./breadcrumb-item.scss";
@@ -73,6 +72,7 @@
 @import "./transfer.scss";
 @import "./tree.scss";
 @import "./upload.scss";
+@import "./avatar.scss";
 @import "./badge.scss";
 @import "./button.scss";
 @import "./container.scss";
