+++ packages/element3/src/components/Avatar/index.js
@@ -1,8 +1,8 @@
-import ElAvatar from './Avatar.vue'
+import ElAvatar from './src/Avatar.vue'
 
 /* istanbul ignore next */
 ElAvatar.install = function (app) {
   app.component(ElAvatar.name, ElAvatar)
 }
 
-export default ElAvatar
+export { ElAvatar }
