+++ packages/element3/src/index.js
@@ -45,7 +45,7 @@ import ElProgress from '../packages/progress'
 import ElTree from '../packages/tree'
 import ElPagination from '../packages/pagination'
 import { ElBadge } from './components/Badge'
-import ElAvatar from '../packages/avatar'
+import { ElAvatar } from './components/Avatar'
 // Notice
 import ElAlert from '../packages/alert'
